Add pull request template
GitHub automatically looks for certain files in the .github directory of the repository, including pull request templates.
If there is a file named PULL_REQUEST_TEMPLATE.md in the .github directory, it will be used as the default pull request template for your repository.
When someone opens a new pull request, they will see the contents of this file as the default description.